Module aws_sdk_ssmcontacts::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

data_encryption_exception

See DataEncryptionException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AcceptPageError

AccessDeniedException

You don't have sufficient access to perform this operation.

ActivateContactChannelError

ConflictException

Updating or deleting a resource causes an inconsistent state.

CreateContactChannelError

CreateContactError

DataEncryptionException

The operation failed to due an encryption key error.

DeactivateContactChannelError

DeleteContactChannelError

DeleteContactError

DescribeEngagementError

DescribePageError

GetContactChannelError

GetContactError

GetContactPolicyError

InternalServerException

Unexpected error occurred while

ListContactChannelsError

ListContactsError

ListEngagementsError

ListPageReceiptsError

ListPagesByContactError

ListPagesByEngagementError

ListTagsForResourceError

PutContactPolicyError

ResourceNotFoundException

Request references a resource that doesn't exist.

SendActivationCodeError

ServiceQuotaExceededException

Request would cause a service quota to be exceeded.

StartEngagementError

StopEngagementError

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateContactChannelError

UpdateContactError

ValidationException

The input fails to satisfy the constraints specified by an Amazon Web Services

Enums

AcceptPageErrorKind

ActivateContactChannelErrorKind

CreateContactChannelErrorKind

CreateContactErrorKind

DeactivateContactChannelErrorKind

DeleteContactChannelErrorKind

DeleteContactErrorKind

DescribeEngagementErrorKind

DescribePageErrorKind

GetContactChannelErrorKind

GetContactErrorKind

GetContactPolicyErrorKind

ListContactChannelsErrorKind

ListContactsErrorKind

ListEngagementsErrorKind

ListPageReceiptsErrorKind

ListPagesByContactErrorKind

ListPagesByEngagementErrorKind

ListTagsForResourceErrorKind

PutContactPolicyErrorKind

SendActivationCodeErrorKind

StartEngagementErrorKind

StopEngagementErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateContactChannelErrorKind

UpdateContactErrorKind