Module aws_sdk_ssmcontacts::model

Modules

channel_target_info

See ChannelTargetInfo

contact

See Contact

contact_channel

See ContactChannel

contact_channel_address

See ContactChannelAddress

contact_target_info

See ContactTargetInfo

engagement

See Engagement

page

See Page

plan

See Plan

receipt

See Receipt

stage

See Stage

tag

See Tag

target

See Target

time_range

See TimeRange

validation_exception_field

See ValidationExceptionField

Structs

ChannelTargetInfo

Information about the contact channel that Incident Manager uses to engage the

Contact

A personal contact or escalation plan that Incident Manager engages during an

ContactChannel

The method that Incident Manager uses to engage a contact.

ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

ContactTargetInfo

The contact that Incident Manager is engaging during an incident.

Engagement

Incident Manager reaching out to a contact or escalation plan to engage contact during an

Page

Incident Manager engaging a contact's contact channel.

Plan

The stages that an escalation plan or engagement plan engages contacts and contact

Receipt

Records events during an engagement.

Stage

A set amount of time that an escalation plan or engagement plan engages the specified

Tag

A container of a key-value name pair.

Target

The contact or contact channel that's being engaged.

TimeRange

A range of between two set times

ValidationExceptionField

Provides information about which field caused the exception.

Enums

AcceptCodeValidation

AcceptType

ActivationStatus

ChannelType

ContactType

ReceiptType

ValidationExceptionReason