Module aws_sdk_ssmcontacts::operation

Structs

AcceptPage

Used to acknowledge an engagement to a contact channel during an incident.

ActivateContactChannel

Activates a contact's contact channel. Incident Manager can't engage a contact until the

CreateContact

Contacts are either the contacts that Incident Manager engages during an incident or the

CreateContactChannel

A contact channel is the method that Incident Manager uses to engage your contact.

DeactivateContactChannel

To no longer receive Incident Manager engagements to a contact channel, you can deactivate

DeleteContact

To remove a contact from Incident Manager, you can delete the contact. Deleting a contact

DeleteContactChannel

To no longer receive engagements on a contact channel, you can delete the channel from a

DescribeEngagement

Incident Manager uses engagements to engage contacts and escalation plans during an incident.

DescribePage

Lists details of the engagement to a contact channel.

GetContact

Retrieves information about the specified contact or escalation plan.

GetContactChannel

List details about a specific contact channel.

GetContactPolicy

Retrieves the resource policies attached to the specified contact or escalation

ListContactChannels

Lists all contact channels for the specified contact.

ListContacts

Lists all contacts and escalation plans in Incident Manager.

ListEngagements

Lists all engagements that have happened in an incident.

ListPageReceipts

Lists all of the engagements to contact channels that have been acknowledged.

ListPagesByContact

Lists the engagements to a contact's contact channels.

ListPagesByEngagement

Lists the engagements to contact channels that occurred by engaging a contact.

ListTagsForResource

Lists the tags of an escalation plan or contact.

PutContactPolicy

Adds a resource to the specified contact or escalation plan.

SendActivationCode

Sends an activation code to a contact channel. The contact can use this code to activate

StartEngagement

Starts an engagement to a contact or escalation plan. The engagement engages each

StopEngagement

Stops an engagement before it finishes the final stage of the escalation plan or

TagResource

Tags a contact or escalation plan. You can tag only contacts and escalation plans in the

UntagResource

Removes tags from the specified resource.

UpdateContact

Updates the contact or escalation plan specified.

UpdateContactChannel

Updates a contact's contact channel.